# src/proxy-ui-api/src/main/resources/openapi-definition.yaml
openapi: 3.0.0
servers:
  - description: X-Road API Design
    url: https://virtserver.swaggerhub.com/Gofore/X-Road/1.0.0
info:
  description: X-Road UI Based API
  version: "1.0.0"
  title: X-Road UI Based API
  contact:
    email: lauri.koutaniemi@gofore.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
...
paths:
...
  /clients:
    get: # ok
      tags:
        - security server
      summary: get security server clients
      operationId: getClients
      description: SS administrator views the clients of the security server.
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: list of clients
          content:
            application/json:
              schema:
                type: array
                uniqueItems: true
                items:
                  $ref: '#/components/schemas/Client'
        '400':
          description: request was invalid
        '401':
          description: authentication credentials are missing
        '403':
          description: request has been refused
        '404':
          description: resource requested does not exists
        '406':
          description: request specified an invalid format
        '500':
          description: internal server error
    post: # ok
      tags:
        - security server
      summary: adds new client for security server
      operationId: addClient
      description: Adds an item to the system
      security:
        - ApiKeyAuth: []
      responses:
        '201':
          description: item created
        '400':
          description: request was invalid
        '401':
          description: authentication credentials are missing
        '403':
          description: request has been refused
        '404':
          description: resource requested does not exists
        '406':
          description: request specified an invalid format
        '409':
          description: an existing item already exists
        '500':
          description: internal server error
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Client'
        description: client to add
  /clients/{id}:
    get: # ok
      tags:
        - security server
      summary: get security server client information
      operationId: getClient
      description: SS administrator views the client details of the security server.
      parameters:
        - in: path
          name: id
          description: id of the client
          required: true
          schema:
            type: string
            format: text
            minLength: 1
            maxLength: 1023
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Client'
        '400':
          description: request was invalid
        '401':
          description: authentication credentials are missing
        '403':
          description: request has been refused
        '404':
          description: resource requested does not exists
        '406':
          description: request specified an invalid format
        '500':
          description: internal server error
    put: # ok
      tags:
        - security server
      summary: update security server client information
      operationId: updateClient
      description: SS administrator updates the client information.
      parameters:
        - in: path
          name: id
          description: id of the client
          required: true
          schema:
            type: string
            format: text
            minLength: 1
            maxLength: 1023
        - in: query
          name: connection_type
          description: server connection type
          required: true
          schema:
            $ref: '#/components/schemas/ConnectionType'
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: item modified
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Client'
        '400':
          description: request was invalid
        '401':
          description: authentication credentials are missing
        '403':
          description: request has been refused
        '404':
          description: resource requested does not exists
        '406':
          description: request specified an invalid format
        '500':
          description: internal server error
    delete: # ok
      tags:
        - security server
      summary: delete security server client
      operationId: deleteClient
      description: SS administrator deletes the client of the security server.
      parameters:
        - in: path
          name: id
          description: id of the client
          required: true
          schema:
            type: string
            format: text
            minLength: 1
            maxLength: 1023
      security:
        - ApiKeyAuth: []
      responses:
        '204':
          description: deletion was successful
        '400':
          description: request was invalid
        '401':
          description: authentication credentials are missing
        '403':
          description: request has been refused
        '404':
          description: resource requested does not exists
        '500':
          description: internal server error
  /clients/{id}/certificates:
    get: # ok
      tags:
        - security server
      summary: get security server client certificates information
      operationId: getClientCertificates
      description: SS administrator views the certificates for the client.
      parameters:
        - in: path
          name: id
          description: id of the client
          required: true
          schema:
            type: string
            format: text
            minLength: 1
            maxLength: 1023
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: list of certificates
          content:
            application/json:
              schema:
                type: array
                uniqueItems: true
                items:
                  $ref: '#/components/schemas/Certificate'
        '400':
          description: request was invalid
        '401':
          description: authentication credentials are missing
        '403':
          description: request has been refused
        '404':
          description: resource requested does not exists
        '406':
          description: request specified an invalid format
        '500':
          description: internal server error
...
  /clients/{id}/tlscertificates:
    get: # ok
      tags:
        - security server
      summary: get security server client TLS certificates information
      operationId: getClientTlsCertificates
      description: SS administrator views the TLS certificates for the client.
      parameters:
        - in: path
          name: id
          description: id of the client
          required: true
          schema:
            type: string
            format: text
            minLength: 1
            maxLength: 1023
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: list of tls certificates
          content:
            application/json:
              schema:
                type: array
                uniqueItems: true
                items:
                  $ref: '#/components/schemas/Certificate'
        '400':
          description: request was invalid
        '401':
          description: authentication credentials are missing
        '403':
          description: request has been refused
        '404':
          description: resource requested does not exists
        '406':
          description: request specified an invalid format
        '500':
          description: internal server error
    post: # ok
      tags:
        - security server
      summary: add new certificate for the security server client
      operationId: addClientTlsCertificate
      description: SS administrator adds a new certificate for the client.
      parameters:
        - in: path
          name: id
          description: id of the client
          required: true
          schema:
            type: string
            format: text
            minLength: 1
            maxLength: 1023
        - in: query
          name: file
          description: import file
          required: true
          schema:
            type: string
            format: binary
      security:
        - ApiKeyAuth: []
      responses:
        '201':
          description: item created
        '400':
          description: request was invalid
        '401':
          description: authentication credentials are missing
        '403':
          description: request has been refused
        '404':
          description: resource requested does not exists
        '406':
          description: request specified an invalid format
        '409':
          description: an existing item already exists
        '500':
          description: internal server error
      requestBody:
        content:
          application/*:
            schema:
              type: string
              format: binary
        description: certificate to add
  /clients/{id}/tlscertificates/{hash}:
    get: # ok
      tags:
        - security server
      summary: get TLS certificate
      operationId: getClientTlsCertificate
      description: SS administrator gets the TLS certificate for the selected client.
      parameters:
        - in: path
          name: id
          description: id of the client
          required: true
          schema:
            type: string
            format: text
            minLength: 1
            maxLength: 1023
        - in: path
          name: hash
          description: SHA-1 hash of the certificate
          required: true
          schema:
            type: string
            format: text
            minLength: 1
            maxLength: 255
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Certificate'
        '400':
          description: request was invalid
        '401':
          description: authentication credentials are missing
        '403':
          description: request has been refused
        '404':
          description: resource requested does not exists
        '406':
          description: request specified an invalid format
        '500':
          description: internal server error
    delete: # ok
      tags:
        - security server
      summary: delete certificate
      operationId: deleteClientTlsCertificate
      description: SS administrator deletes the certificate from selected client.
      parameters:
        - in: path
          name: id
          description: id of the client
          required: true
          schema:
            type: string
            format: text
            minLength: 1
            maxLength: 1023
        - in: path
          name: hash
          description: SHA-1 hash of the certificate
          required: true
          schema:
            type: string
            format: text
            minLength: 1
            maxLength: 255
      security:
        - ApiKeyAuth: []
      responses:
        '204':
          description: deletion was successful
        '400':
          description: request was invalid
        '401':
          description: authentication credentials are missing
        '403':
          description: request has been refused
        '404':
          description: resource requested does not exists
        '500':
          description: internal server error
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: XROAD-API-KEY
    CookieAuth:
      type: apiKey
      in: cookie
      name: JSESSIONID
  schemas:
...
    Certificate: # ok
      required:
      - issuer_distinguished_name
      - issuer_common_name
      - subject_distinguished_name
      - subject_common_name
      - not_before
      - not_after
      - serial
      - version
      - signature_algorithm
      - signature
      - public_key_algorithm
      - rsa_public_key_modulus
      - rsa_public_key_exponent
      - hash
      - key_usages
      - state
      type: object
      properties:
        issuer_distinguished_name:
          type: string
          format: text
          example: issuer123
          description: certificate issuer distinguished name
          minLength: 1
          maxLength: 255
        issuer_common_name:
          type: string
          format: text
          example: domain.com
          description: certificate issuer common name
          minLength: 1
          maxLength: 255
        subject_distinguished_name:
          type: string
          format: text
          example: subject123
          description: certificate subject distinguished name
          minLength: 1
          maxLength: 255
        subject_common_name:
          type: string
          format: text
          example: domain.com
          description: certificate subject common name
          minLength: 1
          maxLength: 255
        not_before:
          type: string
          format: date-time
          example: '2018-12-15T00:00:00.001Z'
          description: certificate validity not before
          minLength: 24
          maxLength: 24
        not_after:
          type: string
          format: date-time
          example: '2018-12-15T00:00:00.001Z'
          description: certificate validity not after
          minLength: 24
          maxLength: 24
        serial:
          type: string
          format: text
          example: '123456789'
          description: serial number
          minLength: 1
          maxLength: 255
        version:
          type: integer
          format: int32
          example: 3
          description: version
        signature_algorithm:
          type: string
          format: sha-256
          example: sha256WithRSAEncryption
          description: certificate signature algorithm
          minLength: 1
          maxLength: 255
        signature:
          type: string
          format: text
          example: '30af2fdc1780...'
          description: hex encoded certificate signature
          minLength: 1
          maxLength: 255
        public_key_algorithm:
          type: string
          format: sha-256
          example: 'sha256WithRSAEncryption'
          description: certificate public key algorithm
          minLength: 1
          maxLength: 255
        rsa_public_key_modulus:
          type: string
          format: hex
          example: c44421d601...
          description: hex encoded RSA public key modulus (if RSA key)
          minLength: 1
          maxLength: 1000
        rsa_public_key_exponent:
          type: integer
          format: int32
          example: 65537
          description: RSA public key exponent (if RSA key) as an integer
          minLength: 1
          maxLength: 255
        hash:
          type: string
          format: text
          example: 1234567890ABCDEF
          description: certificate SHA-1 hash
          minLength: 1
          maxLength: 255
        key_usages:
            type: array
            items:
              $ref: '#/components/schemas/KeyUsage'
        state:
          $ref: '#/components/schemas/State'
    Client: # ok
      type: object
      required:
        - id
        - member_name
        - member_class
        - member_code
        - subsystem_code
        - connection_type
        - status
      properties:
        id:
          type: string
          format: text
          description: <instance_id>:<member_class>:<member_code>:<subsystem>(optional)
          example: FI:GOV:123:ABC
          minLength: 1
          maxLength: 1023
          readOnly: true
        member_name:
          type: string
          format: text
          example: FI
          minLength: 1
          maxLength: 255
        member_class:
          type: string
          format: text
          example: GOV
          minLength: 1
          maxLength: 255
        member_code:
          type: string
          format: text
          example: 123
          minLength: 1
          maxLength: 255
        subsystem_code:
          type: string
          format: text
          example: ABC
          minLength: 1
          maxLength: 255
        connection_type:
          $ref: '#/components/schemas/ConnectionType'
        status:
          $ref: '#/components/schemas/ClientStatus'
    ClientStatus: # ok
      type: string
      format: enum
      example: REGISTERED
      enum:
        - REGISTERED
        - SAVED
        - GLOBAL_ERROR
        - REGISTRATION_IN_PROGRESS
        - DELETION_IN_PROGRESS
    ConnectionType: # ok
      type: string
      format: enum
      example: HTTP
      default: HTTP
      enum:
        - HTTP
        - HTTPS
        - HTTPS_NO_AUTH
...
    KeyUsage: # ok
      type: string
      format: enum
      example: NON_REPUDIATION
      description: certificate key usage
      enum:
        - DIGITAL_SIGNATURE
        - NON_REPUDIATION
        - KEY_ENCIPHERMENT
        - DATA_ENCIPHERMENT
        - KEY_AGREEMENT
        - KEY_CERT_SIGN
        - CRL_SIGN
        - ENCIPHER_ONLY
        - DECIPHER_ONLY
...
    State: # ok
      type: string
      format: enum
      example: IN_USE
      description: certificate state
      enum:
        - IN_USE
        - DISABLED
...
security:
  - ApiKeyAuth: []